refactor(link-wizard): extract CandidateCard component, move LinkWizardStep helpers below return
- CandidateCards mapped candidate JSX inline; extracted to a CandidateCard
  component for reuse/readability
- LinkWizardStep's helper functions were defined before the return
  statement; moved below it per the repo's function-hoisting convention
